Default Bifrost config to merged detectors
